requirements.txt
flask>=2.3.0
python-docx>=0.8.11
markdown>=3.4.0
Werkzeug>=2.3.0
